% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my.loess01.R
\name{my.loess1}
\alias{my.loess1}
\title{My Loess Function Title}
\usage{
my.loess1(formula, data, weights, subset, na.action, model = FALSE,
  span = 0.75, enp.target, degree = 2L, parametric = FALSE,
  drop.square = FALSE, normalize = TRUE, family = c("gaussian",
  "symmetric"), method = c("loess", "model.frame"),
  control = loess.control(...), ...)
}
\arguments{
\item{formula}{FORMULA DESCRIPTION. DO THIS FOR EACH ONE}
}
\description{
My Loess details
}
\examples{
my.loess1()
}
\author{
Xiaosu Tong
}

