% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my.predloess.R
\name{my.predict.loess}
\alias{my.predict.loess}
\title{My Loess Function Title}
\usage{
my.predict.loess(object, newdata = NULL, se = FALSE, na.action = na.pass,
  ...)
}
\arguments{
\item{formula}{FORMULA DESCRIPTION. DO THIS FOR EACH ONE}
}
\description{
My Loess details
}
\examples{
  a <- my.loess1(dist ~ speed, cars, control = loess.control(surface = "direct"))
  str(a)
  my.predict.loess(a, data.frame(speed = seq(5, 30, 1)))
}
\author{
Xiaosu Tong
}

