% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my.predloess.R
\name{my.predLoess}
\alias{my.predLoess}
\title{My Loess Function Title}
\usage{
my.predLoess(y, x, newx, s, weights, robust, span, degree, normalize,
  parametric, drop.square, surface, cell, family, kd, divisor, se = FALSE,
  distance)
}
\arguments{
\item{formula}{FORMULA DESCRIPTION. DO THIS FOR EACH ONE}
}
\description{
My Loess details
}
\examples{
my.loess1()
}
\author{
Xiaosu Tong
}

