% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my.loess01.R
\name{my.simple1}
\alias{my.simple1}
\title{My Loess Function Title}
\usage{
my.simple1(y, x, weights, span = 0.75, degree = 2L, parametric = FALSE,
  drop.square = FALSE, normalize = TRUE, statistics = "approximate",
  surface = "interpolate", cell = 0.2, iterations = 1L,
  trace.hat = "exact")
}
\arguments{
\item{formula}{FORMULA DESCRIPTION. DO THIS FOR EACH ONE}
}
\description{
My Loess details
}
\examples{
my.loess1()
}
\author{
Xiaosu Tong
}

