% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newPredLoess.R
\name{newPredLoess}
\alias{newPredLoess}
\title{Spatially Local Polynomial Regression Prediction}
\usage{
newPredLoess(y, x, newx, s, weights, robust, span, degree, normalize,
  parametric, drop.square, surface, cell, family, kd, divisor, se = FALSE,
  distance)
}
\arguments{
\item{y}{the response variable in the form of vector}

\item{x}{a data.frame of the original design matrix}

\item{newx}{a data.frame containing the new locations for prediction}

\item{s}{standard error from the orginial local fit. If se = FALSE, s will not be useful.}

\item{weights}{optional initial weights for the original data set. Get from spaloess$weights.}

\item{robust}{robust factor to each location of original data set, get from spaloess$pars$robust.}

\item{span}{The parameter alpha which controls the portion of data points used in the local fit.}

\item{degree}{The degree of the polynomials to be used, normally 1 or 2. (Degree 0 is also allowed, but see 
the 'Note'.)}

\item{normalize}{Should the predictors be normalized to a common scale if there is more than one?  The 
normalization used is to set the 10% trimmed standard deviation to one. Set to false for 
"Latlong" distance.}

\item{parametric}{should any terms be fitted globally rather than locally? Terms can be specified by name, 
number or as a logical vector of the same length as the number of predictors.}

\item{drop.square}{For fits with more than one predictor and 'degree = 2', should the quadratic term be dropped 
for particular predictors?  Terms are specified in the same way as for 'parametric'.}

\item{surface}{should the fitted surface be computed exactly or via interpolation from a kd tree? 
can be either "interpolate" or "direct", default is "interpolate".}

\item{cell}{if interpolation is used this controls the accuracy of the approximation via the maximum 
number of points in a cell in the kd tree. Cells with more than 'floor(n*span*cell)' points
are subdivided.}

\item{family}{If 'gaussian' fitting is by least-squares, and if 'symmetric' a re-descending M estimator is 
used with Tukey's bi-weight function.}

\item{kd}{a list object which contains all information about the kd-tree built based on original data 
set.}

\item{divisor}{the scaler for the normalization. x/divisor if normalize = TRUE.}

\item{se}{should standard errors be computed? Default is FALSE}

\item{distance}{Specify the method of distance calculation. Options: "Euclid", or "Latlong" which is for great 
circle distance}

\item{...}{arguments passed to or from other methods.}
}
\description{
The second layer of the prediction of Spatial locally weighted regression. It calls 
the computation engine functions wrote in C and Fortran, Concretely, the "loess_dfit"
and "loess_ifit". Notice, it has to be called by predloess function, do not call 
newPredLoess directly.
}
\author{
Xiaosu Tong, based on 'loess' function of B. D. Ripley, and 'cloess' package of Cleveland,
   Grosse and Shyu.
}

