% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predloess.R
\name{predloess}
\alias{predloess}
\title{Spatially Local Polynomial Regression Prediction}
\usage{
predloess(object, newdata = NULL, se = FALSE, na.action = na.pass, ...)
}
\arguments{
\item{object}{an object fitted by 'spaloess'.}

\item{newdata}{an optional data frame in which to look for variables with which to predict, or a 
matrix or vector containing exactly the variables needs for prediction.  If missing, 
the original data points are used.}

\item{se}{should standard errors be computed? Default is FALSE}

\item{na.action}{function determining what should be done with missing values in data frame 'newdata'.
The default is to predict 'NA'.}

\item{...}{arguments passed to or from other methods.}
}
\description{
The first layer of the prediction of Spatial locally weighted regression. Mainly used as
prediction function for the NAs in the original data set.
}
\details{
This is the first layer of prediction function of spatial locally weigted regression.
    In the spaloess function, NA will be removed from the fitting. By passing the spaloess
    object and NA observations to predloess, predction at the locations of NA is carried out.

    When the fit was made using 'surface = "interpolate"' (the default), 'predloess' 
    will not extrapolate - so points outside an axis-aligned hypercube enclosing the 
    original data will have missing ('NA') predictions and standard errors.
}
\examples{
    set.seed(66)
    x1 <- rnorm(100, mean=-100, sd=10)
    x2 <- rnorm(100, mean=38, sd=4)
    y <- 0.1*x1 + 1*x2 - 10 + rnorm(100, 0, 1.3); y[1:2] <- NA
    testdata <- data.frame(LON = x1, LAT = x2, tmax = y)
    cars.lo <- spaloess(tmax ~ LON + LAT, testdata, distance = "Latlong")
}
\author{
Xiaosu Tong, based on 'loess' function of B. D. Ripley, and 'cloess' package of Cleveland,
    Grosse and Shyu.
}

