% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spaloess.R
\name{spaloess}
\alias{spaloess}
\title{Spatially Local Polynomial Regression Fitting}
\usage{
spaloess(formula, data, weights, subset, na.action, model = FALSE,
  span = 0.05, enp.target, degree = 2L, parametric = FALSE,
  distance = "Latlong", drop.square = FALSE, normalize = FALSE,
  family = c("gaussian", "symmetric"), method = c("loess", "model.frame"),
  control = loess.control(...), ...)
}
\arguments{
\item{formula}{FORMULA DESCRIPTION. DO THIS FOR EACH ONE}

\item{data}{an optional data from, list or environment containing the variables in the model. If 
not found in 'data', the variables are taken from 'environment', typically the environment 
from which 'loess' is called.}

\item{weights}{optional weights for each case}

\item{subset}{an optional specification of a subset of the data to be used}

\item{na.action}{the action to be taken with missing values in the response or predictors.  The default is 
given by 'getOption("na.action")'.}

\item{model}{Should the model frame be returned?}

\item{span}{The parameter alpha which controls the degree of smoothing.}

\item{enp.target}{An alternative way to specify 'span', as the approximate equivalent number of parameters to 
be used.}

\item{degree}{The degree of the polynomials to be used, normally 1 or 2. (Degree 0 is also allowed, but see 
the ‘Note’.)}

\item{parametric}{should any terms be fitted globally rather than locally? Terms can be specified by name, 
number or as a logical vector of the same length as the number of predictors.}

\item{distance}{Options: "Euclid", "Mahal", or "Latlong" which is for great circle distance}

\item{drop.square}{For fits with more than one predictor and 'degree = 2', should the quadratic term be dropped 
for particular predictors?  Terms are specified in the same way as for 'parametric'.}

\item{normalize}{Should the predictors be normalized to a common scale if there is more than one?  The 
normalization used is to set the 10% trimmed standard deviation to one. Set to false for 
"Latlong" distance function, and true for "Mahal" distance.}

\item{family}{If 'gaussian' fitting is by least-squares, and if 'symmetric' a re-descending M estimator is 
used with Tukey's bi-weight function.}

\item{method}{Fit the model or just extract the model frame.}

\item{control}{control parameters: see 'loess.control'.}
}
\description{
The first layer of the Spatial locally weighted regression, using local fitting with different 
type of distance calculation.
}
\details{
This spaloess function is the first wrapper of the spatial loess fitting procedure. It checks
    all the validity of all input arguments, and formats arguments like drop,square, parametric.
    Also generate other important arguments, like iteration, and pass all arguments into the 
    second wrapper function: newsimpleLoess
}
\examples{
    cars.lo <- loess(dist ~ speed, cars, distance = "Mahal")
}
\author{
Xiaosu Tong, based on 'loess' function of B. D. Ripley, and 'cloess' package of Cleveland,
    Grosse and Shyu.
}

